create styles.js for styled components
