+++ src/components/HomePage/HomePage.js
@@ -16,66 +16,8 @@ import {
   faInstagram,
 } from '@fortawesome/free-brands-svg-icons';
 import * as barberActions from '../../actions/Actions';
+import {Title, Service} from '../../styles'
 
-const color = ({ currentColor }) => currentColor;
-
-const Service = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  text-align: center;
-  overflow: hidden;
-  font-family: 'Lato', black;
-  .image-place {
-    align-items: center;
-    background-color: ${color};
-    width: 100px;
-    height: 100px;
-    border-radius: 50%;
-    box-sizing: border-box;
-    text-align: center;
-    margin-left: auto;
-    margin-right: auto;
-
-    img {
-      height: 100%;
-    }
-  }
-  .barber-name {
-    text-align: center;
-    p {
-      color: black;
-      border-bottom: 1px dotted gray;
-      width: 50%;
-      margin: 0 auto;
-      font-weight: 700;
-    }
-  }
-  svg {
-    margin-left: 5px;
-    cursor: pointer;
-  }
-
-  .details {
-    color: gray;
-    font-weight: 500;
-  }
-
-  .social {
-    color: gray;
-  }
-`;
-const Title = styled.div`
-  font-family: 'Lato', black;
-  text-align: center;
-
-  h3 {
-    font-weight: 900;
-  }
-  p {
-    color: gray;
-  }
-`;
 
 function HomePage({ ...props }) {
   const settings = {
